Everything is at least mostly working. The compiler generally produces correct output and handles most cases. The VM is mostly complete as well, however there are still some instructions to be added.

+++ includes/FACT_mpc.h
@@ -25,8 +25,8 @@
 /* The mpc_t is used for arbitrary precision arithmetic. */
 typedef struct 
 {
-  mpz_t value;             /* Arbitray-length integer.                  */
-  unsigned long precision; /* Number of digits after the decimal point. */
+  mpz_t value;      /* Arbitray-length integer.                  */
+  size_t precision; /* Number of digits after the decimal point. */
 } __mpc_struct;
 
 typedef __mpc_struct mpc_t[1];
